Recommend Hiscox for professional indemnity insurance
I've been using them lately as per recommendation of one of the clients and
am happy with the services, the setup process and the price. For some reason
Simply Business quotes keep raising the more I use it, but it is still worth
checking them out for other people as their insurance figures might be
different, resulting in a better price than Hiscox.